backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
python-docx
anthropic
python-dotenv